Update packages to latest
Removes es6-promise, which is not necessary for local development (we're all on a modern version of node).
Moves the sass file generation to a separate function, the current IIFE method is not clear.
Update autoprefixer's browserlist to use the core WordPress configuration
Add a default task, and map grunt in npm scripts